Don't over estimate the number of units to invalidate
Previously we were always invalidating at least one row when moving right, even if we  weren't moving up.  This was to work around the artifact that moving right would sometimes invalidate a partial row.  I've isolated the cause of that artifact and we now work around it, removing the need to always invalidate one row/column.